Addind edit and delete for users and clients

+++ src/client/client.service.ts
@@ -40,54 +40,68 @@ export class ClientService {
         };
     }
 
-    async findOne(id: string): Promise<{ statusCode: number, message: string, client: Client }> {
-        const client = await this.clientRepository.findOne({
-            where: { id }
-        });
+    async findOneById(id: string): Promise<{ statusCode: number, message: string, client: Client }> {
+            const client = await this.clientRepository.findOne({ where: { id } });
 
-        if (!client) {
-            throw new NotFoundException('Cliente no encontrado');
+            if (!client) {
+                throw new NotFoundException(`Client with ID ${id} not found`);
+            }
+            return {
+                statusCode: HttpStatus.OK,
+                message: 'Client found successfully',
+                client: client
+            };
         }
 
+
+    async updateClient(id: string, updateData: Partial<Client>): Promise<{ statusCode: number, message: string }> {
+        const clientData = await this.findOneById(id);
+
+        if (updateData.email) {
+            const existingClient = await this.clientRepository.findOne({
+                where: [
+                    { email: updateData.email, id: Not(id) },
+                    { phone: updateData.phone, id: Not(id) }
+                ]
+            });
+            if (existingClient) {
+                throw new ConflictException('Ya existe un cliente con ese email o teléfono');
+            }
+        }
+
+        Object.assign(clientData.client, updateData);
+        await this.clientRepository.save(clientData.client);
         return {
             statusCode: HttpStatus.OK,
-            message: 'Cliente encontrado correctamente',
-            client: client
+            message: 'Cliente actualizado correctamente'
         };
     }
 
-    // async update(id: string, updateData: Partial<Client>): Promise<{ statusCode: number, message: string, client: Client }> {
-    //     const client = await this.findOne(id);
-
-    //     if (updateData.email || updateData.phone) {
-    //         const existingClient = await this.clientRepository.findOne({
-    //             where: [
-    //                 { email: updateData.email, id: Not(id) },
-    //                 { phone: updateData.phone, id: Not(id) }
-    //             ]
-    //         });
-
-    //         if (existingClient) {
-    //             throw new ConflictException('Ya existe un cliente con ese email o teléfono');
-    //         }
-    //     }
-
-    //     Object.assign(client, updateData);
-    //     await this.clientRepository.save(client);
-    //     return {
-    //         statusCode: HttpStatus.OK,
-    //         message: 'Cliente actualizado correctamente',
-    //         client: client
-    //     };
-    // }
-
-    async remove(id: string): Promise<{ statusCode: number, message: string }> {
-        const clientData = await this.findOne(id);
+    async removeClient(id: string): Promise<{ statusCode: number, message: string }> {
+        const clientData = await this.findOneById(id);
+        
+        if (!clientData) {
+            throw new NotFoundException(`Client with ID ${id} not found`);
+        }
         await this.clientRepository.remove(clientData.client);
         return {
             statusCode: HttpStatus.OK,
-            message: 'Cliente eliminado correctamente'
+            message: `Client with ID ${id} deleted successfully`
         };
     }
+
+    // async removeUser(id: string): Promise<{ statusCode: number, message: string }> {
+    //     const user = await this.findOneById(id);
+
+    //     if (!user) {
+    //         throw new NotFoundException(`User with ID ${id} not found`);
+    //     }
+
+    //     await this.userRepository.remove(user);
+    //     return {
+    //         statusCode: 200,
+    //         message: `User with ID ${id} deleted successfully`
+    //     };
+    // }
 
 }
